Fix interceptor discovery for GET parameters
I am currently trying to integrate your (btw. amazing and amazingly simple) tool into a large code base for mocking a large quantity of API calls. I discovered that for some reason some of the calls were not intercepted and could track it down to the interceptor failing in case there are GET parameters in the request URL and the allow_unmocked option is set. I haven't had time yet to check further but simply removing this bit makes all tests work for me here...